requirements.txt
redis>=5.0.8
mysql-connector-python>=9.1.0
requests>=2.32.3
beautifulsoup4>=4.12.3
python-dotenv>=1.2.1
numpy>=1.26.0,<2.0.0
scikit-learn>=1.5.0
cython>=3.0.11
pyyaml>=6.0.2
openai>=1.40.0
langchain>=1.0.0
langchain-classic>=1.0.0
langchain-community>=0.4.0
langchain-openai>=0.2.1
langgraph>=0.2.0
chromadb>=0.5.0
langchain-chroma>=1.0.0
jieba>=0.42.1
sentence-transformers>=3.0.0
matplotlib>=3.8.0